+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetop.py
@@ -100,7 +100,7 @@ class CreateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-08-19T10:58:01.279Z")
+    updated: Optional[datetime] = parse_datetime("2026-08-19T12:36:36.185Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")