+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsejsonschema.py
@@ -42,9 +42,9 @@ class EvaluatorResponseJSONSchema(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-08-19T10:58:05.439Z"
+    created: Optional[str] = "2026-08-19T12:36:38.391Z"
 
-    updated: Optional[str] = "2026-08-19T10:58:05.439Z"
+    updated: Optional[str] = "2026-08-19T12:36:38.391Z"
 
     updated_by_id: OptionalNullable[str] = UNSET